+++ Zup/app/src/main/java/school/sptech/zup/domain/model/FeedRequest.kt
@@ -1,13 +1,11 @@
 package school.sptech.zup.domain.model
 
 
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
 import java.io.Serializable
 
 
 data class FeedRequest(
-    val id: Long,
+    val id: Integer,
     val titulo: String,
     val descricao: String,
     val link: String,
